+++ src/components/MultiSelect/multiSelectChangelogData.ts
@@ -10,6 +10,15 @@ import { ChangelogData } from "../../utils/ComponentChangelogTable";
 
 export const changelogData: ChangelogData[] = [
   {
+    date: "Prerelease",
+    version: "Prerelease",
+    type: "Bug Fix",
+    affects: ["Styles"],
+    notes: [
+      "Fixes the vertical alignment of the label for the `inline` variant.",
+    ],
+  },
+  {
     date: "2025-01-30",
     version: "3.5.3",
     type: "Update",